Refactor typography previews
use the typography preview in style panels

Pass through the font size
