+++ packages/edit-site/src/components/global-styles/style-variations-container.js
@@ -47,6 +47,7 @@ export default function StyleVariationsContainer() {
 							label={ variation?.title }
 							withHoverView
 							isFocused={ isFocused }
+							variation={ variation }
 						/>
 					) }
 				</Variation>
