+++ packages/edit-site/src/components/global-styles/variations/style.scss
@@ -37,7 +37,7 @@
 	}
 }
 
-.edit-site-global-styles-variations__type-preview {
+.edit-site-global-styles_preview-typography {
 	font-size: 22px;
 	line-height: 44px;
 	text-align: center;
